Resolving ticket 247

+++ src/main/java/ui/interactivesimulation/JFrameInteractiveSimulation.java
@@ -1984,7 +1984,7 @@ public class JFrameInteractiveSimulation extends JFrame implements ActionListene
                                 String val = node0.getTextContent();
                                 TraceManager.addDev("Sim time=" + val);
                                 int valueCycle = Integer.decode(val);
-                                long timeP = (long)((long)(valueCycle) * 1000 / frequency);
+                                long timeP = ((long)(valueCycle) * 1000 / frequency);
                                 val = formatString(val);
                                 String timePS = formatString(""+timeP);
                                 val = val + " cycles / " + timePS  + " ns";
